requirements.txt
flask
flask-cors
tensorflow>=2.9
pillow
numpy
matplotlib
scikit-learn
ortools
opencv-python
